requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.37.12
botocore==1.37.12
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels_redis==4.2.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.8.0rc2
dj-database-url==2.3.0
dj-rest-auth==7.0.0
Django==5.1.3
django-admin-interface==0.29.4
django-allauth==65.2.0
django-anymail==12.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-storages==1.14.5
django-taggit==6.1.0
django-timezone-field==7.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
fido2==1.1.3
flower==2.0.1
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
humanize==4.11.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.2
msgpack==1.1.0
newsapi-python==0.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
s3transfer==0.11.4
service-identity==24.2.0
setuptools==75.5.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
text-unidecode==1.3
tornado==6.4.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.0
whitenoise==6.9.0
zope.interface==7.2
